requirements.txt
alembic==1.12.0
asgiref==3.7.2
Django==4.2.5
django-ninja==0.22.2
greenlet==2.0.2
Mako==1.2.4
MarkupSafe==2.1.3
pydantic==1.10.13
SQLAlchemy==2.0.21
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
